+++ src/authentication/infrastructure/http/create-user.dto.ts
@@ -1,4 +1,9 @@
+import { ApiProperty } from '@nestjs/swagger';
+
 export class CreateUserRequestDto {
+  @ApiProperty({ example: 'porter@torre-norte.edificios' })
   email!: string;
+
+  @ApiProperty({ example: 'Carlos Porter' })
   displayName!: string;
 }